server/requirements.txt
alembic==1.6.5
autopep8==1.5.7
blinker==1.4
click==8.0.1
colorama==0.4.4
config==0.5.0.post0
dbr==8.6
Flask==2.0.1
Flask-Caching==1.10.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
protobuf==3.17.3
pycodestyle==2.7.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
qrcode==7.3.1
six==1.16.0
SQLAlchemy==1.4.23
toml==0.10.2
Werkzeug==2.0.1
